Reorganize concurrent set and st_table optimization
Move the code to dedicated modules to improve readability.  Also renamed
USDT names for consistency.
